Backend/classifier/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==7.1.2
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.68.0
h11==0.14.0
idna==3.10
motor==3.6.1
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
PyJWT==2.8.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.14.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.13.4
pandas==2.2.3
seaborn==0.13.2
numpy==2.2.2
matplotlib==3.10.0
plotly==5.24.1
tenacity==9.0.
kaleido==0.2.1